Separate out log building in retry plugin
